+++ src/Microsoft.Diagnostics.Runtime/src/Linux/ElfSymbolTable.cs
@@ -7,12 +7,12 @@ namespace Microsoft.Diagnostics.Runtime.Linux
     internal class ElfSymbolTable
     {
         private readonly Reader _reader;
-        private readonly long _address;
+        private readonly ulong _address;
         private readonly bool _is64Bit;
         private readonly ElfStringTable _stringTable;
-        private readonly int _symSize;
+        private readonly uint _symSize;
 
-        public ElfSymbolTable(Reader reader, bool is64Bit, long address, ElfStringTable stringTable)
+        public ElfSymbolTable(Reader reader, bool is64Bit, ulong address, ElfStringTable stringTable)
         {
             _reader = reader;
             _address = address;
@@ -21,12 +21,12 @@ namespace Microsoft.Diagnostics.Runtime.Linux
             _stringTable = stringTable;
         }
 
-        public ElfSymbol GetSymbol(int index)
+        public ElfSymbol GetSymbol(uint index)
         {
             if (_is64Bit)
             {
                 ElfSymbol64 s = _reader.Read<ElfSymbol64>(_address + index * _symSize);
-                string name = _stringTable.GetStringAtIndex((int)s.Name);
+                string name = _stringTable.GetStringAtIndex(s.Name);
                 ElfSymbolBind b = (ElfSymbolBind)(s.Info >> 4);
                 ElfSymbolType t = (ElfSymbolType)(s.Info & 0xF);
                 return new ElfSymbol(name, b, t, s.Value, (long)s.Size);
@@ -34,7 +34,7 @@ namespace Microsoft.Diagnostics.Runtime.Linux
             else
             {
                 ElfSymbol32 s = _reader.Read<ElfSymbol32>(_address + index * _symSize);
-                string name = _stringTable.GetStringAtIndex((int)s.Name);
+                string name = _stringTable.GetStringAtIndex(s.Name);
                 ElfSymbolBind b = (ElfSymbolBind)(s.Info >> 4);
                 ElfSymbolType t = (ElfSymbolType)(s.Info & 0xF);
                 return new ElfSymbol(name, b, t, (long)s.Value, s.Size);
